Moved UI for menus to own package
