+++ app/src/main/java/de/ironjan/mensaupb/activities/Menus.java
@@ -35,8 +35,8 @@ import de.ironjan.mensaupb.BuildConfig;
 import de.ironjan.mensaupb.MensaUpbApplication;
 import de.ironjan.mensaupb.monitoring.MonitoringConstants;
 import de.ironjan.mensaupb.R;
-import de.ironjan.mensaupb.adapters.WeekdayHelper;
-import de.ironjan.mensaupb.adapters.WeekdayPagerAdapter;
+import de.ironjan.mensaupb.menus_ui.WeekdayHelper;
+import de.ironjan.mensaupb.menus_ui.WeekdayPagerAdapter;
 import de.ironjan.mensaupb.prefs.InternalKeyValueStore_;
 import de.ironjan.mensaupb.stw.Restaurant;
 import de.ironjan.mensaupb.stw.opening_times.OpeningTimesKeeper;
