+++ app/src/main/java/de/ironjan/mensaupb/menus_ui/DateHelper.java
@@ -6,7 +6,7 @@ import java.util.*;
 /**
  * TODO merge with WeekDayHelper
  *
- * @see de.ironjan.mensaupb.adapters.WeekdayHelper
+ * @see WeekdayHelper
  */
 public class DateHelper {
