+++ app/src/main/java/de/ironjan/mensaupb/menus_ui/MenuListingAdapter.java
@@ -1,4 +1,4 @@
-package de.ironjan.mensaupb.adapters;
+package de.ironjan.mensaupb.menus_ui;
 
 
 import android.content.Context;
@@ -19,7 +19,7 @@ import de.ironjan.mensaupb.prefs.AllergenFilterPrefs_;
 import de.ironjan.mensaupb.stw.rest_api.StwMenu;
 import de.ironjan.mensaupb.sync.MenuContentProvider;
 import de.ironjan.mensaupb.menus_ui.MenuListingHeaderView;
-import de.ironjan.mensaupb.views.MenuListingHeaderView_;
+import de.ironjan.mensaupb.menus_ui.MenuListingHeaderView_;
 import se.emilsjolander.stickylistheaders.StickyListHeadersAdapter;
 
 /**
