+++ app/src/main/java/de/ironjan/mensaupb/menus_ui/MenuListingFragment.java
@@ -21,9 +21,6 @@ import org.slf4j.LoggerFactory;
 import de.ironjan.mensaupb.BuildConfig;
 import de.ironjan.mensaupb.R;
 import de.ironjan.mensaupb.activities.MenusNavigationCallback;
-import de.ironjan.mensaupb.adapters.MenuDetailViewBinder;
-import de.ironjan.mensaupb.adapters.MenuListingAdapter;
-import de.ironjan.mensaupb.fragments.MenuListingFragment_;
 import de.ironjan.mensaupb.prefs.InternalKeyValueStore_;
 import de.ironjan.mensaupb.stw.opening_times.OpeningTimesKeeper;
 import de.ironjan.mensaupb.sync.AccountCreator;
